refactor: extract TestingLaunchAtLoginService.swift from LaunchAtLoginService.swift
Extracts the `TestingLaunchAtLoginService` test double (~15 lines)
from LaunchAtLoginService.swift into its own file, matching the
existing pattern of pairing a Testing* stub with its production peer.

Byte-identical move of lines 98-112. New file only needs Foundation
(no ServiceManagement — the testing double never touches SMAppService).
LaunchAtLoginService.swift keeps its Foundation + ServiceManagement.

LaunchAtLoginService.swift: 112 → 96 lines. Tests: 667.
