+++ Sources/BugNarrator/Services/LaunchAtLoginService.swift
@@ -94,19 +94,3 @@ struct SystemLaunchAtLoginService: LaunchAtLoginControlling {
         return currentStatus()
     }
 }
-
-struct TestingLaunchAtLoginService: LaunchAtLoginControlling {
-    let status: LaunchAtLoginStatus
-
-    init(status: LaunchAtLoginStatus = .disabled) {
-        self.status = status
-    }
-
-    func currentStatus() -> LaunchAtLoginStatus {
-        status
-    }
-
-    func setEnabled(_ enabled: Bool) throws -> LaunchAtLoginStatus {
-        enabled ? .enabled : .disabled
-    }
-}
